Work on x64 build: webp and snapshots are fixed, Win64 Release is added, some optimizations are enabled.

+++ source/common.c
@@ -469,7 +469,7 @@ next
   1..x : used,next is used
 */
 
-void hash_add(hashtable_t *t,const char *s,int s_sz,int key,int mode)
+void hash_add(hashtable_t *t,const char *s,int s_sz,intptr_t key,int mode)
 {
     hashitem_t *cur;
     char buf[BUFLEN];
@@ -536,7 +536,7 @@ void hash_add(hashtable_t *t,const char *s,int s_sz,int key,int mode)
     if(previ>=0)(&t->items[previ])->next=curi;
 }
 
-int hash_find(hashtable_t *t,char *s,int sz,int *isfound)
+intptr_t hash_find(hashtable_t *t,char *s,int sz,int *isfound)
 {
     hashitem_t *cur;
     char buf[BUFLEN];
@@ -603,7 +603,7 @@ int hash_find(hashtable_t *t,char *s,int sz,int *isfound)
     return 0;
 }
 
-int hash_find_str(hashtable_t *t,char *s)
+intptr_t hash_find_str(hashtable_t *t,char *s)
 {
     int b;
     return hash_find(t,s,strlen(s),&b);
@@ -646,7 +646,7 @@ int hash_findnext_b(hashtable_t *t,int *isfound)
     return 0;
 }
 
-int hash_findnext(hashtable_t *t)
+intptr_t hash_findnext(hashtable_t *t)
 {
     int b;
     return hash_findnext_b(t,&b);
